+++ bwplus_sounds_merlin/$PREFIX$
@@ -1,0 +1,1 @@
+y\BWplus\addons\sounds_merlin
