+++ bwplus_sounds_weapons/$PREFIX$
@@ -1,0 +1,1 @@
+y\BWplus\addons\sounds_weapons
