+++ bwplus_sounds_weapons/script_component.hpp
@@ -1,0 +1,6 @@
+#define COMPONENT sounds_weapons
+
+#include "\y\BWplus\addons\main\script_mod.hpp"
+
+
+#include "\y\BWplus\addons\main\script_macros.hpp"
